backend/requirements.txt
django-filter==22.1
djoser==2.1.0
drf-extra-fields==3.4.1
flake8==5.0.4
isort==5.11.5
pillow==9.4.0
pip-chill==1.0.1
postgres==4.0
psycopg2-binary==2.9.5
pydocstyle==5.0.0
gunicorn==20.0.4
